Move validators to new namespace, update DI scan logic
Moved CancelBookingCommandValidator and RejectBookingCommandValidator to the Bookings.Validators namespace and updated their usings. Modified ArchitecturalDiscoveryHelper to exclude CommandHandler and Validator classes from self-registration during service scanning.
