+++ src/Modules/Bookings/Application/Bookings/Validators/CancelBookingCommandValidator.cs
@@ -1,6 +1,7 @@
 using FluentValidation;
+using MeAjudaAi.Modules.Bookings.Application.Bookings.Commands;
 
-namespace MeAjudaAi.Modules.Bookings.Application.Bookings.Commands;
+namespace MeAjudaAi.Modules.Bookings.Application.Bookings.Validators;
 
 public class CancelBookingCommandValidator : AbstractValidator<CancelBookingCommand>
 {
